Simplify release workflows
I noticed that the workflow reference to the pre-release-checks was
still pointing to the add-release-workflows branch.

Then I realised that we could probably store that workflow in our
central .github repo, so I moved it there.

I also realised that the references to the workflows that are in this
repo can use a path syntax instead.