ruby: Remove redundant test against cck
The main purpose of the formatter is to write the given template
correctly. This is already covered by the non-cck tests. As such the
dependency on the CCK doesn't add anything.

The cck incidentally does check that the empty report is written
correctly. Which can be done with a fairly minimal test.